heroku_deployed_citi_bike_app/requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
Brotli==1.0.9
cffi==1.15.0
click==8.0.3
cramjam==2.5.0
dash==2.0.0
dash-core-components==2.0.0
dash-deck==0.0.1
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fastparquet==0.7.2
Flask==2.0.2
Flask-Compress==1.10.1
fsspec==2021.11.1
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.22.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
plotly==5.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
Send2Trash==1.8.0
six==1.16.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
thrift==0.15.0
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
